requirements.txt
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.7.1
bing-images==0.0.3
Bootstrap-Flask==1.0.10
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.4
Flask==1.0.2
Flask-WTF==0.14.2
future==0.17.1
idna==2.8
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.2.0
playsound==1.2.2
pycodestyle==2.6.0
pydub==0.23.1
pylint==2.6.0
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==1.9.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.24.2
Werkzeug==0.15.3
wiktionaryparser==0.0.97
wrapt==1.12.1
WTForms==2.2.1